src/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.9
django-cors-headers==4.5.0
django-filter==24.3
django-jalali-date==1.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
Khayyam==3.0.17
mssql-django==1.5
paho-mqtt==2.1.0
pika==1.3.2
pillow==11.0.0
PyJWT==2.9.0
pyodbc==5.2.0
pytz==2024.2
requests==2.32.3
setuptools==75.2.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
python-dotenv==1.1.0
